refactor: Split multiclusterhub_controller.go into logical components
Split the large 2027-line multiclusterhub_controller.go file into 10
smaller, logically organized files to improve maintainability and readability:

- reconcile.go: Main Reconcile function
- sts.go: STS (Security Token Service) functionality
- components.go: Component lifecycle management
- templates.go: Template and resource management
- internal_hub_components.go: InternalHubComponent CR management
- infrastructure.go: Cluster infrastructure utilities
- lifecycle.go: Installation and finalization
- config.go: Configuration and overrides
- storage.go: Storage class management
- setup.go: Controller setup and status management

All functions maintain their original functionality with proper imports
and package structure. The original file now contains only struct
definitions, constants, and RBAC annotations.
